Update comments & split components
