+++ packages/types/src/index.ts
@@ -3,6 +3,7 @@ export * from './discord/applicationRoleConnectionMetadata.js'
 export * from './discord/auditLog.js'
 export * from './discord/autoModeration.js'
 export * from './discord/channel.js'
+export * from './discord/components.js'
 export * from './discord/emoji.js'
 export * from './discord/entitlement.js'
 export * from './discord/gateway.js'
